[PersonInGroup] Refacto RestController et Service Personne, bascule de tous les traitements métier sur le ServicePersonne
