+++ src/main/spa/src/app/components/person/person.component.ts
@@ -1,4 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
+import { Person } from 'src/app/shared/models/person';
+import { AuthService } from 'src/app/shared/services/auth.service';
+import { PersonService } from 'src/app/shared/services/person.service';
+import { UserService } from 'src/app/shared/services/user.service';
 
 @Component({
   selector: 'app-person',
@@ -7,9 +12,35 @@ import { Component, OnInit } from '@angular/core';
 })
 export class PersonComponent implements OnInit {
 
-  constructor() { }
+  public person?: Person | null;
+
+  constructor(
+    private _personService : PersonService,
+    private _route: ActivatedRoute,
+    private _router : Router,
+    private _userService: UserService,
+    private _authService : AuthService,
+
+  ) { }
+
+  getPrenomNom(){
+    return this.person?.prenom +" "+ this.person?.nom?.toUpperCase();
+  }
 
   ngOnInit(): void {
+    this._userService.updateToken();
+    this._route.paramMap.subscribe(param => {
+      const personId = this._userService.decodedToken.id;
+      if(personId != null && personId > 0) {
+        this._personService.getPersonById(personId)
+        .subscribe((p: Person | null) => {
+          this.person = p;
+        });
+      }
+      else {
+        this._router.navigateByUrl("/home");
+      }
+    });
   }
 
 }
